+++ R/plsda.R
@@ -1,4 +1,8 @@
-plsda <- function(Xr, Yr, Xu, Yu = NULL, ncomp, algo = pls.kernel, da = dalm, ...) {
+plsda <- function(Xr, Yr, Xu, Yu = NULL, ncomp, 
+                  algo = NULL, da = dalm, ...) {
+  
+  if(is.null(algo))
+    algo <- pls.kernel
   
   dots <- list(...)
   namdot <- names(dots)